Move tests for spect_spec.rb to integration_spec.rb
Due to the recursive nature of Spect it was difficult to test
Spect#conform! and Spect#validate! as pure unit tests (i.e not
testing any other classes, such as Keys) and it made tests difficult
to understand. As a result, Spect#conform! and Spect#validate are
tested an integration tests.